Update pioneer_pub.cpp
The wheel velocities have been updated to float type from int type. corresponding changes to be done in the pioneerROS.ttt file as well

+++ pioneer_pub.cpp
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "rclcpp/rclcpp.hpp"
-#include "std_msgs/msg/int32.hpp"
+#include "std_msgs/msg/Float32.hpp"
 using namespace std::chrono_literals;
 
 class SpeedPublisher : public rclcpp::Node
@@ -12,27 +12,27 @@ class SpeedPublisher : public rclcpp::Node
 public:
     SpeedPublisher() : Node("pioneer_pub")
     {
-    leftpublisher_ = this->create_publisher<std_msgs::msg::Int32>("leftMotorSpeed",10);
-    rightpublisher_ = this->create_publisher<std_msgs::msg::Int32>("rightMotorSpeed",10);
+    leftpublisher_ = this->create_publisher<std_msgs::msg::Float32>("leftMotorSpeed",10);
+    rightpublisher_ = this->create_publisher<std_msgs::msg::Float32>("rightMotorSpeed",10);
     timer_ = this->create_wall_timer(500ms, std::bind(&SpeedPublisher::timer_callback, this));
     }
     
     private:
     void timer_callback()
     {
-    auto message_left = std_msgs::msg::Int32();
-    message_left.data = 5;
+    auto message_left = std_msgs::msg::Float32();
+    message_left.data = 5.0;
     RCLCPP_INFO(this->get_logger(), "Publishing left wheel speed: '%d'", message_left.data);
     leftpublisher_->publish(message_left);
     
-    auto message_right = std_msgs::msg::Int32();
-    message_right.data = 5;
+    auto message_right = std_msgs::msg::Float32();
+    message_right.data = 5.0;
     RCLCPP_INFO(this->get_logger(), "Publishing right wheel speed: '%d'",message_right.data);
     rightpublisher_->publish(message_right);
     }
     rclcpp::TimerBase::SharedPtr timer_;
-    rclcpp::Publisher<std_msgs::msg::Int32>::SharedPtr leftpublisher_;
-    rclcpp::Publisher<std_msgs::msg::Int32>::SharedPtr rightpublisher_;
+    rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr leftpublisher_;
+    rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr rightpublisher_;
     
     };
